+++ tests/test_function/test_parse_file.py
@@ -22,6 +22,19 @@ def test_parse_load_file_as_base64() -> None:
     assert_that(result).is_equal_to(expected_result)
 
 
+def test_parse_load_json_content() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "load_json_content"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BICEP_PARSER.parse(file_path=file_path)
+
+    # then
+    assert_that(result).is_equal_to(expected_result)
+
+
 def test_parse_load_text_content() -> None:
     # given
     sub_dir_path = EXAMPLES_DIR / "load_text_content"